DAVE-205 Gesamtauswertung Ausgabe Excel (#162)
* excel wird prototypisch erzeugt

* wip

* wip

* move pom dependency

* regroup code

* move method of spreadsheetservice

* move method of spreadsheetservice

* refactor code

* refactor code

* refactor code

* add javadoc

* streamify nested for loop

* streamify nested for loop

* rename method names

* refactor code

* refactor code

* remove unused import

* add mst id to response info

* add mst id to response info

* rename class and attribute

* rename class and attribute

* refactor code

* refactor code

* remove unused docs

* remove unused docs

* make code more resilient

* adapt to new api docs for geodata eai

* add todo

* umgestellt auf neue Datenstruktur

* generierung des files eingebaut

* sort results

* sort results

* renaming

* add Tests for AuswertungService

* logs entfernt

* add javadoc

* Reviewanmerkungen umgesetzt

* Reviewanmerkungen umgesetzt

* spotless

---------

+++ src/main/java/de/muenchen/dave/domain/dtos/laden/messwerte/LadeProcessedMesswerteDTO.java
@@ -24,6 +24,6 @@ public class LadeProcessedMesswerteDTO implements Serializable {
 
     BelastungsplanMessquerschnitteDTO belastungsplanMessquerschnitte;
 
-    private Integer requestedMeasuringDays;
+    private Long requestedMeasuringDays;
     private Integer includedMeasuringDays;
 }
